requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorful==0.5.4
fake-useragent==0.1.11
idna==2.10
numpy==1.20.1
prettyprinter==0.18.0
Pygments==2.8.1
redis==3.5.3
requests==2.25.1
selenium==3.141.0
soupsieve==2.2
urllib3==1.26.4

python-dotenv~=0.20.0
python-dateutil~=2.8.2